Refactoring: Split reading and writing of translation files
